requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
scipy>=1.7.0

# Data processing
pymysql>=1.0.0
sqlalchemy>=1.4.0
redis>=4.0.0

# Machine learning
xgboost>=1.5.0
lightgbm>=3.3.0
catboost>=1.0.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0
dash>=2.0.0

# Web framework
fastapi>=0.70.0
uvicorn>=0.15.0
streamlit>=1.12.0

# MCP and AI integration
openai>=1.0.0
anthropic>=0.5.0
zhipuai>=1.0.0
dashscope>=1.0.0

# Async and networking
aiohttp>=3.8.0
asyncio-mqtt>=0.11.0
websockets>=10.0

# Configuration and utilities
pyyaml>=6.0
click>=8.0.0
python-dotenv>=0.19.0
pydantic>=1.8.0

# Logging and monitoring
loguru>=0.6.0
prometheus-client>=0.12.0

# Development and testing
pytest>=6.2.0
pytest-asyncio>=0.18.0
pytest-cov>=3.0.0
black>=22.0.0
flake8>=4.0.0
mypy>=0.910

# Jupyter and examples
jupyter>=1.0.0
ipykernel>=6.0.0
notebook>=6.4.0

# Docker and deployment
gunicorn>=20.1.0
psutil>=5.8.0